requirements.lock.txt
annotated-types==0.7.0
anyio==4.9.0
authlib==1.6.0
bandit==1.8.5
boolean-py==5.0
cachecontrol==0.14.3
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
coverage==7.9.1
cryptography==45.0.4
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distlib==0.3.9
dparse==0.6.4
fastapi==0.115.13
filelock==3.16.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
joblib==1.5.1
license-expression==30.4.1
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
msgpack==1.1.1
mypy==1.16.1
mypy-extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
packageurl-python==0.17.1
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
pip==25.1.1
pip-api==0.0.34
pip-audit==2.9.0
pip-requirements-parser==32.0.1
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==4.2.0
psutil==6.1.1
py-serializable==2.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
ruff==0.12.0
safety==3.5.2
safety-schemas==0.0.14
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
stevedore==5.4.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
